(PE-2793) Make Jetty client-auth configurable
This commit adds the ability to configure the client-auth Jetty setting
in the webserver section of the config file.  The setting can have a
value of `need` (server requires that the client provide a certificate
and will validate the certificate), `want` (server asks the client for a
certificate and will validate the certificate if one is provided), or
`none` (server will not ask the client for a certificate).  The default
value, if not specified in the config, is the same as what was used in
previous releases, `need`.